float64 instead float for cdouble convertion

+++ src/allegro5/altime.nim
@@ -10,13 +10,13 @@ proc al_rest(seconds: cdouble): void
 proc al_init_timeout(timeout: var AllegroTimeout, seconds: cdouble): void
 {.pop.}
 
-proc time*(): float =
+proc time*(): float64 =
   return al_get_time().cdouble
 
-proc rest*(seconds: float): void =
+proc rest*(seconds: float64): void =
   al_rest(seconds.cdouble)
 
-proc newAllegroTimeout*(seconds: float): AllegroTimeout =
+proc newAllegroTimeout*(seconds: float64): AllegroTimeout =
   var timeout = AllegroTimeout()
   al_init_timeout(timeout, seconds.cdouble)
   return timeout
